Perform hit detection in physical LED space
Previously hit detection was done in virtual LED space (i.e. 0-1000) but
this caused some weirdness, especially with a small number of LED. For
example you could end up in a situation where visually the edge of an
attack totally hits an enemy, but the enemy does not die...because in
virtual space the attack didn't quite reach the enemy even though they
are aliased to the same LED on the strip. This was confusing. Now all
hit detection is done in the actual LED space, so what you see is what
you get.